Int to float
Changed from grabbing an int to a float to take into account for half
cups, etc

+++ Lab2/cups_to_ounces_modified.py
@@ -8,7 +8,7 @@ def main():
     intro()
     
     #KV Find out cups needed from the user
-    cups_needed = int(input('Enter the number of cups: '))
+    cups_needed = float(input('Enter the number of cups: '))
     
     #KV Call function to Convert from cups to ounces
     cups_to_ounces(cups_needed)
